requirement.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
confusable-homoglyphs==3.2.0
Cython==0.29.24
decorator==5.0.9
Django==1.5.12
django-crispy-forms==1.12.0
django-crispy-forms-registration==0.1.3
django-extensions==3.1.3
django-registration==3.2
dyNET==2.1.1
idna==3.2
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
nagisa==0.2.7
numpy==1.21.1
oauthlib==3.1.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
tweepy==3.10.0
urllib3==1.26.6
wcwidth==0.2.5
